Slight alteration to Limit Orders
The limit price can now be equal to the high or low of the day.  This was changed because sometimes, when checking a stop loss, the open may be the high/low of the day and therefore is not strictly between the high and low. This should still allow the stop loss to be activated.