Changed order of trade functions
If the `trade_open` function will execute, then  the `trade_every_day_open` will execute after this. The reason is so stop-losses can be checked after orders have been placed at the open, as this is what the `trade_every_day_open` was designed to do. In future, I will be making it optional which functions will run when.