Added take-profit and made new optimisation option
There is now code for check_take_profit but it has not been fully tested. I have also given the option for 'combination' optimisation, which cycles through every possible combination of the opt_params, or 'fixed' optimisation, which requires each paramter to have the same number of options and then use the first one of all of them, then the second of all of them and so on.